+++ morib/src/main/java/org/morib/server/domain/task/application/FetchTaskService.java
@@ -1,4 +1,4 @@
-package org.morib.server.api.timerView.service.fetch.task;
+package org.morib.server.domain.task.application;
 
 import java.util.Set;
 import org.morib.server.domain.category.infra.Category;
@@ -7,5 +7,5 @@ import org.morib.server.domain.task.infra.Task;
 public interface FetchTaskService {
     void fetch();
 
-    Task fetchByTaskIdInCategories(Set<Category> categories, Long taskId);
+    Task fetchById(Long taskId);
 }
