+++ morib/src/main/java/org/morib/server/domain/task/application/FetchTaskServiceImpl.java
@@ -1,0 +1,28 @@
+package org.morib.server.domain.task.application;
+
+import lombok.RequiredArgsConstructor;
+import org.morib.server.domain.task.infra.Task;
+import org.springframework.stereotype.Service;
+
+@Service
+@RequiredArgsConstructor
+public class FetchTaskServiceImpl implements FetchTaskService {
+
+    private final TaskManager taskManager;
+
+    @Override
+    public void fetch() {
+        //timer가 멈췄을때 계산해주는 역할을 해야함!
+
+    }
+
+    @Override
+    public Task fetchById(Long taskId) {
+
+
+        return null;
+    }
+
+
+
+}
