+++ morib/src/main/java/org/morib/server/domain/task/application/TaskManager.java
@@ -1,4 +1,4 @@
-package org.morib.server.domain.task;
+package org.morib.server.domain.task.application;
 
 import lombok.RequiredArgsConstructor;
 import org.morib.server.domain.task.infra.Task;
